Flytter hjelpefunksjon til EndringstidspunktUtils
